Replaces ncmpcpp with mpc usage. Creates .xinitrc service to notice mpd seeks
ncmpcpp and mpc are both command line interfaces to mpd. ncmpcpp is interactive,
whereas mpc is command line. Earlier I had used xmonad hotkeys and a script
$HOME/bin/musicPlayback to change songs and display the current song; however,
it did not detect natural song changes in the mpd playlist itself.

  mpc currenty --wait

blocks waiting for song changes and when used in an infinite loop is a more
complete solution. So now the 'musicPlayback' script simply displays the current
song.